Update Kuksa databroker configuration scheme
Rework Kuksa databroker configuration to also read a configuration
yaml file from /etc/xdg/AGL/ics-homescreen.yaml at runtime to allow
overriding the built-in default configuration.  The latter is left
as is with the addition of an explicit disabling of TLS, to still
match the internal test setup at ICS.  The AGL build will supply
a runtime configuration that enables TLS and an authorization token.

Additionally:
- the ValClientHelper constructor has been changed to take the
  stub and authorization token as arguments to allow authorization
  with the VAL API requests.  The channel argument has been
  removed since it is not used, only the stub is.
- the apparently unused vss_provider.dart source file has been
  removed to avoid potential confusion.

Bug-AGL: SPEC-4969

Change-Id: Ic7b867eaf565ef22f2e70df0afe7b5272b604d04
Signed-off-by: Scott Murray <scott.murray@konsulko.com>