Connect audio settings button
Rework to connect the audio settings button on the media pages
to the appropriate settings page, and have the back button go
back to the expected previous page.  To enable this, the global
AppState provider was reworked to track the previous page and add
explicit update and back member functions.

Bug-AGL: SPEC-5030

Change-Id: I5858d1b1bf511a184b6538b2ce8c183b00c24fc6
Signed-off-by: Scott Murray <scott.murray@konsulko.com>