Add support for CI screenshot testing
Rework the high-level App & AppView widgets a bit to allow
conditionally displaying a simple fixed pattern if the
HOMESCREEN_DEMO_CI environment variable is set to anything
other than "0".  The pattern displayed matches what the
Qt homescreen shows for this to allow the CI test to be
common.

Bug-AGL: SPEC-5203

Change-Id: I36ed77c91304e3b6cdbce6a7350831d52d20d378
Signed-off-by: Scott Murray <scott.murray@konsulko.com>